Conflicts:
	configure.py

+++ src/canon_perftest.cc
@@ -27,7 +27,7 @@ int main() {
   string err;
 
   char buf[200];
-  int len = strlen(kPath);
+  size_t len = strlen(kPath);
   strcpy(buf, kPath);
 
   for (int j = 0; j < 5; ++j) {
